app/django/meridien/requirements.txt
asgiref==3.2.7
astroid==2.4.1
certifi==2020.4.5.1
Django==3.0.7
django-cors-headers==3.3.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
factory-boy==2.12.0
Faker==4.1.0
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
psycopg2==2.8.5
PyJWT==1.7.1
pylint==2.5.2
python-dateutil==2.8.1
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
text-unidecode==1.3
toml==0.10.1
wrapt==1.12.1
